move get_case_sync_updates to shared module
